requirements.txt
mwparserfromhell==0.5.4
PyYAML==5.1.1
